Refactor `tag_li.dart` to make it possible to customize marker widget
